added alternative circumcenter algo, minimal alpha algo
