+++ canhydro/geometry.py
@@ -1,0 +1,311 @@
+
+from typing import Union
+import math
+
+import matplotlib.pyplot as plt 
+import numpy as np
+import geopandas as geo
+
+from shapely.geometry import Polygon
+from scipy.spatial import Delaunay, distance
+from shapely.ops import polygonize, unary_union as union
+from shapely.geometry import Polygon
+
+from scipy.linalg import lu_factor, lu_solve 
+import scipy.sparse 
+
+from canhydro.global_vars import log
+from canhydro.DataClasses import coord_list
+
+
+def circumcenter_lapack(points: coord_list) -> np.ndarray:
+    """
+    Calculate the circumcenter of a set of points relative to simplex
+    https://en.wikipedia.org/wiki/Polarization_identity
+    """
+    points = np.asarray(points)
+    rows, _ = points.shape
+    A = np.bmat(obj =[[2 * np.dot(points, points.T), np.ones((rows, 1))], [np.ones((1, rows)), np.zeros((1, 1))]])
+    b = np.hstack((np.sum(points * points, axis=1),np.ones((1))))
+    sir_c = np.linalg.solve(A, b)[:-1]
+    return sir_c
+
+def circumcenter_lu_factor(points: coord_list) -> np.ndarray:
+    """
+    Calculate the circumcenter of a set of points relative to simplex
+    Theoretically less efficient than LAPACK (O(n^3) + O(n^2) v. O(n^2)) when a given set of equations (e.g. set of points, A )
+        only needs to be solved for a single vector (e.g b)
+    """
+    points = np.asarray(points)
+    rows, columns = points.shape
+    A = np.bmat(obj =[[2 * np.dot(points, points.T), np.ones((rows, 1))], [np.ones((1, rows)), np.zeros((1, 1))]])
+    b = np.hstack((np.sum(points * points, axis=1),
+                   np.ones((1))))
+    lu, piv = lu_factor(A)
+    sir_c = lu_solve((lu, piv), b)
+    return sir_c
+
+def circumradius(points: coord_list) -> float:
+    """
+    Calculte the radius of the circle in which the given polygon may be inscribed 
+    """
+    points = np.asarray(points)
+    return np.linalg.norm(points[0, :] - np.dot(circumcenter_LAPACK(points), points))
+
+def simplices(points: coord_list)->coord_list:
+    """
+        Yeilds simpicies and radius
+    """
+    coords = np.asarray(points)
+    tri = Delaunay(coords)
+
+    for simplex in tri.simplices:
+        simplex_points = coords[simplex]
+        try:
+            yield simplex, circumradius(simplex_points)
+        except np.linalg.LinAlgError:
+            log.Warning('Singleton returned - likely caused by all points '
+                          'lying in an N-1 space.')
+
+def maximal_alpha(boundary_points: coord_list, union_poly: Polygon)-> float:
+    """
+    Finds the minimal alpha shape for the given coord list that still contains the given polygon
+    """
+    upper = 10 # annectotally seems to be plenty high to ensure a discontinuous alpha shape
+    lower = 0
+    candidate = start
+    itterations = 10
+    while runs<= itterations:
+        alpha = (upper-lower)/2
+        hull, _ = concave_hull(boundary_points, union_poly)
+        if hull.contains(Polygon):
+            # The tested curvature is less than or equal to the max curvature for a continuous alpha shape 
+            lower = alpha
+        else:
+            # The tested curvature is greater than or equal to the max curvature for a continuous alpha shape 
+            upper = alpha            
+        runs+=1
+    return alpha
+
+#might eventually be updated to deal with 3d a la https://github.com/bellockk/alphashape/blob/master/alphashape/optimizealpha.py
+def concave_hull(boundary_points, alpha:int= 0):
+    """alpha shape / concave hull
+    Draws a minimal concave polygon with a concavity factor alpha"""
+    
+    if len(boundary_points) < 4:
+        # When you have a triangle, there is no sense in computing an alpha
+        # shape.
+        return geometry.MultiPoint(list(boundary_points)).convex_hull, boundary_points
+
+    def add_edge(edges, edge_points, coords, i, j):
+        # adds a line between points i and j
+        if (i, j) in edges or (j, i) in edges:
+            # already added
+            return
+        edges.add((i, j))
+        edge_points.append(coords[[i, j]])
+
+    coords = np.array(sorted([point.coords[0] for point in boundary_points if len(point.coords)>0]))
+    # Minimal set of triangles with points in set
+
+    edges = set()
+    edge_points = []
+    # loop over triangles:
+    # ia, ib, ic = indices of corner points of the triangle
+    tri = simplices(coords)
+    for (ia, ib, ic), _ in tri:
+        breakpoint()
+        pa = coords[ia]
+        pb = coords[ib]
+        pc = coords[ic]
+
+        # Lengths of sides of triangle
+        a = math.sqrt((pa[0] - pb[0]) ** 2 + (pa[1] - pb[1]) ** 2)
+        b = math.sqrt((pb[0] - pc[0]) ** 2 + (pb[1] - pc[1]) ** 2)
+        c = math.sqrt((pc[0] - pa[0]) ** 2 + (pc[1] - pa[1]) ** 2)
+
+        # Semiperimeter of triangle
+        s = (a + b + c) / 2.0
+
+        # Area of triangle by Heron's formula
+        area = math.sqrt(s * (s - a) * (s - b) * (s - c))
+        circum_r = a * b * c / (4.0 * area)
+
+        # Here's the radius filter.
+        # print circum_r
+        if circum_r < 1.0 / alpha:
+            add_edge(edges, edge_points, coords, ia, ib)
+            add_edge(edges, edge_points, coords, ib, ic)
+            add_edge(edges, edge_points, coords, ic, ia)
+
+    m = geometry.MultiLineString(edge_points)
+    triangles = list(polygonize(m))
+    return unary_union(triangles), edge_points
+
+def union(list):
+    return unary_union(list)
+
+def furthest_point(point, points):
+    furthest_index = distance.cdist([point], points).argmax()
+    return points[furthest_index]
+
+def get_projection(vector:list, magnitude:list, radius:float()):
+    noCirPoints = 360
+    tCir = np.linspace(
+        0, 2 * np.pi, noCirPoints
+    )  # 360 evenly spaced points between 0 - 2pi (radian degrees)
+    a_ortho = np.cos(tCir)  # x coordinates of the points on a circle
+    b_ortho = np.sin(tCir)  # y coordinates of the points on a circle
+    delt_a = magnitude[0]
+    delt_b = magnitude[1]
+    delt_c = magnitude[2]
+    dim_a = vector[0]
+    dim_b = vector[1]
+    dim_c = vector[2]
+    # unit vector at base of cylinder, pointing up cylinder axis
+    vNorm = np.sqrt(delt_a**2 + delt_b**2 + delt_c**2)
+    aV = np.hstack((delt_a, delt_b, delt_c)) / vNorm
+    bV = -aV  # unit vector looking down from top circle (but not translated)
+    # function to find orthgonal vectors
+    oVz = lambda v, a, b: ((-v[0] * a - v[1] * b) / v[2])
+    # initializing min max arrays+
+    min_c = np.zeros_like(delt_c)
+    max_c = np.zeros_like(delt_c)
+    pSV = []
+    projection = {
+                    "polygon": Polygon(),
+                    "base_vector": None,
+                    "anti_vector": None,
+                    "angle": None,
+                    "area": None,
+                }
+    try:
+        # for each cylinder
+        if not np.isnan(dim_a[0]):
+            if np.logical_and(delt_a == 0, delt_b == 0):
+                pX = dim_a[0] + radius * a_ortho
+                pY = dim_b[0] + radius * b_ortho
+                cPS = Polygon(list(zip(pX, pY)))
+                min_c = np.min(dim_c[:])
+                max_c = np.max(dim_c[:])
+            else:
+                # find orthogonal vectors @ endpoints
+                # Identifies corners of projected rectangle
+                aVp1 = np.hstack((aV[1], -aV[0]))
+                aVp2 = np.hstack((-aV[1], aV[0]))
+                bVp1 = np.hstack((bV[1], -bV[0]))
+                bVp2 = np.hstack((-bV[1], bV[0]))
+                aVp1 = aVp1 / np.linalg.norm(aVp1)
+                aVp2 = aVp2 / np.linalg.norm(aVp2)
+                bVp1 = bVp1 / np.linalg.norm(bVp1)
+                bVp2 = bVp2 / np.linalg.norm(bVp2)
+                # from each endpoint, use radius to find vertices of the rectangle
+                x1 = dim_a[0] + radius * aVp1[0]
+                y1 = dim_b[0] + radius * aVp1[1]
+                x2 = dim_a[0] + radius * aVp2[0]
+                y2 = dim_b[0] + radius * aVp2[1]
+                x3 = dim_a[1] + radius * bVp1[0]
+                y3 = dim_b[1] + radius * bVp1[1]
+                x4 = dim_a[1] + radius * bVp2[0]
+                y4 = dim_b[1] + radius * bVp2[1]
+                # calculate set of orthgonal vectors using lambda function
+                # That is 360 orthogonal vectors ending at eqidistant points along
+                # a circle of radius radius with the starting point of our cylinder
+                # at is center
+                ZOrtho = oVz(aV[:], a_ortho, b_ortho)
+                # unit-ify the orthgonal vectors
+                uovd = np.sqrt(a_ortho**2 + b_ortho**2 + ZOrtho**2)
+                # Confounded - why does removing the first three [:,None]'s below lead to non-circular projections
+                # for XZ?
+                uov = (
+                    np.hstack((a_ortho[:, None], b_ortho[:, None], ZOrtho[:, None]))
+                    / uovd[:, None]
+                )
+                # donot re unit-fy, you only want the horizontal component, not the
+                # renormalized horizontal component
+                # using only the X and Y components, find circle coods in plane of
+                # interest
+                xaC = dim_a[0] + uov[:, 0] * radius
+                yaC = dim_b[0] + uov[:, 1] * radius
+                zaC = dim_c[0] + uov[:, 2] * radius
+                xbC = dim_a[1] + uov[:, 0] * radius
+                ybC = dim_b[1] + uov[:, 1] * radius
+                zbC = dim_c[1] + uov[:, 2] * radius
+                min_c = np.min(np.vstack((zaC, zbC)))
+                max_c = np.max(np.vstack((zaC, zbC)))
+                # assymble total package
+                rX = np.vstack((x1, x2, x3, x4))
+                rY = np.vstack((y1, y2, y3, y4))
+                # test for circle parts in polygon
+                try:
+                    c1 = Polygon(
+                        list(
+                            zip(
+                                [0 if np.isnan(x) else x for x in xaC],
+                                [0 if np.isnan(y) else y for y in yaC],
+                            )
+                        )
+                    )
+                    bBox = Polygon(
+                        list(
+                            zip(
+                                [0 if np.isnan(x) else x for x in rX],
+                                [0 if np.isnan(y) else y for y in rY],
+                            )
+                        )
+                    )
+                    c2 = Polygon(
+                        list(
+                            zip(
+                                [0 if np.isnan(x) else x for x in xbC],
+                                [0 if np.isnan(y) else y for y in ybC],
+                            )
+                        )
+                    )
+                    partsPS = [c1, bBox, c2]
+                except:
+                    log.info("Error creating projection polygons")
+                try:
+                    cPS = unary_union(partsPS)
+                except:
+                    print(np.any(np.isnan(xaC)))
+                    log.info("Error unioning projection polygons")
+                    print(yaC)
+                    print(rX)
+                    print(rY)
+                    print(xbC)
+                    print(ybC)
+                # get angle away from plane projected on to
+                run = math.sqrt(delt_b**2 + delt_a**2)
+                rise = delt_c
+                if run == 0:
+                    slope = 1  # straightDown e.g. is in flow
+                else:
+                    slope = rise / run
+                ang = np.arctan(slope)
+                projection = {
+                    "polygon": cPS,
+                    "base_vector": aV,
+                    "anti_vector": bV,
+                    "angle": ang,
+                    "area": cPS.area,
+                }
+                return projection
+        else:
+            log.info(f"dim_a[0] is null, unable to project")
+        return projection
+    except UnboundLocalError:
+        log.info(f"UnboundLocalError: vector : {vector} magnitude: {magnitude} radius: {radius}")
+
+    
+def draw_cyls(collection: Union[list[Polygon],Polygon], colors: list[bool] = [True]):
+    log.info("Plotting cylinder collection")
+    fig, ax = plt.subplots()
+    geoPolys = GeoSeries(collection)
+    colors = ["Blue" if col else "Grey" for col in colors]
+    geoPolys.plot(ax=ax, color=colors)
+    plt.show()
+
+
+def unary_union(polys):
+    return union(polys)
