+++ tests/expected_results_shapes.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+from shapely import Polygon
+
+star_poly =  Polygon([(0.75, 0.5), (1, 0), (0.5, 0.25), (0, 0), (0.25, 0.5), (0, 1), (0.5, 0.75), (1, 1), (0.75, 0.5)])
+
