+++ tests/test_alpha_shape.py
@@ -6,80 +6,138 @@ import itertools
 import shapely
 from alphashape.alphashape import alphashape
 from alphashape import cli
-from canhydro.utils import lam_filter, read_file_names, concave_hull
+from shapely.geometry import Point
+from canhydro.utils import lam_filter, read_file_names
 
+from canhydro.geometry import concave_hull, furthest_point, union
 
+from tests.expected_results_shapes import (
+    star_poly
+)
 class TestAlphashape(unittest.TestCase):
     """Tests for `alphashape` package."""
 
-    def setUp(self):
-        """Set up test fixtures, if any."""
+    # def setUp(self):
+    #     """Set up test fixtures, if any."""
 
-    def tearDown(self):
-        """Tear down test fixtures, if any."""
+    # def tearDown(self):
+    #     """Tear down test fixtures, if any."""
 
-    def test_given_a_point_return_a_point(self):
-        """
-        Given a point, the alphashape function should return the same point
-        """
-        point_a, _ = concave_hull([(0., 0.)], 0)
-        point_c, _ = concave_hull([(1., 0.)], 0)
-        point_b, _ = concave_hull([(0., 1.)], 0)
-        point_d, _ = concave_hull([(0., 0.)], 99)
-        point_e, _ = concave_hull([(1., 0.)], 99)
-        point_f, _ = concave_hull([(0., 1.)], 99)
-        assert shapely.geometry.Point([0., 0.]) == point_a
-        assert shapely.geometry.Point([1., 0.]) == point_c
-        assert shapely.geometry.Point([0., 1.]) == point_b
-        assert shapely.geometry.Point([0., 0.]) == point_d
-        assert shapely.geometry.Point([1., 0.]) == point_e
-        assert shapely.geometry.Point([0., 1.]) == point_f
+    # def test_given_a_point_return_a_point(self):
+    #     """
+    #     Given a point, the alphashape function should return the same point
+    #     """
+    #     point_a, _ = concave_hull([(0., 0.)], 0)
+    #     point_c, _ = concave_hull([(1., 0.)], 0)
+    #     point_b, _ = concave_hull([(0., 1.)], 0)
+    #     point_d, _ = concave_hull([(0., 0.)], 99)
+    #     point_e, _ = concave_hull([(1., 0.)], 99)
+    #     point_f, _ = concave_hull([(0., 1.)], 99)
+    #     assert shapely.geometry.Point([0., 0.]) == point_a
+    #     assert shapely.geometry.Point([1., 0.]) == point_c
+    #     assert shapely.geometry.Point([0., 1.]) == point_b
+    #     assert shapely.geometry.Point([0., 0.]) == point_d
+    #     assert shapely.geometry.Point([1., 0.]) == point_e
+    #     assert shapely.geometry.Point([0., 1.]) == point_f
 
-    def test_given_a_line_with_dupicate_points_return_a_point(self):
-        """
-        Given a line with duplicate points, the alphashape function should
-        return a point
-        """
-        actual, _  = concave_hull([(0., 1.), (0., 1.)], 0)
-        assert shapely.geometry.Point([0., 1.]) == actual
+    # def test_given_a_line_with_dupicate_points_return_a_point(self):
+    #     """
+    #     Given a line with duplicate points, the alphashape function should
+    #     return a point
+    #     """
+    #     actual, _  = concave_hull([(0., 1.), (0., 1.)], 0)
+    #     assert shapely.geometry.Point([0., 1.]) == actual
 
-    def test_given_a_line_with_unique_points_return_a_line(self):
-        """
-        Given a line with unique points, the alphashape function should return
-        the same line
-        """
+    # def test_given_a_line_with_unique_points_return_a_line(self):
+    #     """
+    #     Given a line with unique points, the alphashape function should return
+    #     the same line
+    #     """
         
-        actual_horz, _= concave_hull([(0., 0.), (0., 1.)], 0)
-        actual_diag, _= concave_hull([(1., 0.), (0., 1.)], 0)
-        assert shapely.geometry.LineString([(0., 0.), (0., 1.)]) == actual_horz
-        assert shapely.geometry.LineString([(1., 0.), (0., 1.)]) == actual_diag
+    #     actual_horz, _= concave_hull([(0., 0.), (0., 1.)], 0)
+    #     actual_diag, _= concave_hull([(1., 0.), (0., 1.)], 0)
+    #     assert shapely.geometry.LineString([(0., 0.), (0., 1.)]) == actual_horz
+    #     assert shapely.geometry.LineString([(1., 0.), (0., 1.)]) == actual_diag
 
-    def test_given_a_triangle_with_duplicate_points_returns_a_point(self):
-        """
-        Given a triangle with two unique points, the alphashape function should
-        return a point
-        """
-        dupe_tri, _ = concave_hull([(0., 1.), (0., 1.), (0., 1.)], 0)
-        assert shapely.geometry.Point((0., 1.)) == dupe_tri
+    # def test_given_a_triangle_with_duplicate_points_returns_a_point(self):
+    #     """
+    #     Given a triangle with two unique points, the alphashape function should
+    #     return a point
+    #     """
+    #     dupe_tri, _ = concave_hull([(0., 1.), (0., 1.), (0., 1.)], 0)
+    #     assert shapely.geometry.Point((0., 1.)) == dupe_tri
 
-    def test_given_a_triangle_with_two_duplicate_points_returns_a_line(self):
-        """
-        Given a line with two unique points, the alphashape function should
-        return a line with the unique points
-        """
-        tri_two_dupe,_ = concave_hull([(1., 0.), (0., 1.), (0., 1.)], 0)
-        assert shapely.geometry.LineString([(1., 0.), (0., 1.)]) == tri_two_dupe
+    # def test_given_a_triangle_with_two_duplicate_points_returns_a_line(self):
+    #     """
+    #     Given a line with two unique points, the alphashape function should
+    #     return a line with the unique points
+    #     """
+    #     tri_two_dupe,_ = concave_hull([(1., 0.), (0., 1.), (0., 1.)], 0)
+    #     assert shapely.geometry.LineString([(1., 0.), (0., 1.)]) == tri_two_dupe
 
-    def test_given_a_four_point_polygon_with_small_alpha_return_input(self):
-        """
-        Given a polygon with four points, and an alpha value of zero, return
-        the input as a polygon.
-        """
-        quad_small, _= concave_hull([(0., 0.), (0., 1.), (1., 1.), (1., 0.)], 1.e-9)
-        breakpoint()
-        assert shapely.geometry.Polygon([(0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.)]).equals(quad_small)
+    # def test_given_a_four_point_polygon_with_small_alpha_return_input(self):
+    #     """
+    #     Given a polygon with four points, and an alpha value of zero, return
+    #     the input as a polygon.
+    #     """
+    #     pts = [Point((x,y)) for (x,y) in  [(0., 0.), (0., 1.), (1., 1.), (1., 0.)]]
+    #     quad_small, _= concave_hull(pts, .0001)
+    #     assert shapely.geometry.Polygon([(0., 0.), (0., 1.), (1., 1.), (1., 0.), (0., 0.)]).equals(quad_small)
 
-    def test_3_dimensional_regression_with_dynamic_alpha(self):
+    def test_star_poly(self):
+        """
+        Given a 3-dimensional data set, return an expected set of edges.
+        """
+        coords_2d = sorted([
+            (0., 0.), (0., 0.), (0., 1.),
+            (1., 0.), (1., 1.), (1., 0.),
+            (0., 1.), (1., 1.), (.25, .5),
+            (.5, .25), (.5, .5), (.75, .5),
+            (.5, .75), (.5, .5)
+        ])
+        points_2d = [Point((x,y)) for (x,y) in coords_2d]
+        expected_vertices = [
+            [0., 0.], [0., 0.], [0., 1.],
+            [0., 1.], [1., 0.], [1., 0.],
+            [1., 1.], [1., 1.], [0.25, 0.5],
+            [0.5, 0.25], [0.5, 0.5], [0.5, 0.5],
+            [0.5, 0.75], [0.75, 0.5]
+            ]
+        
+        expected = star_poly
+        results, _ = concave_hull(points_2d,2.1)
+        assert results.contains(expected)
+        assert expected.contains(results)
+        # breakpoint()
+
+    def test_star_poly(self):
+        """
+        Given a 3-dimensional data set, return an expected set of edges.
+        """
+        coords_2d = sorted([
+            (0., 0.), (0., 0.), (0., 1.),
+            (1., 0.), (1., 1.), (1., 0.),
+            (0., 1.), (1., 1.), (.25, .5),
+            (.5, .25), (.5, .5), (.75, .5),
+            (.5, .75), (.5, .5)
+        ])
+        points_2d = [Point((x,y)) for (x,y) in coords_2d]
+        expected_vertices = [
+            [0., 0.], [0., 0.], [0., 1.],
+            [0., 1.], [1., 0.], [1., 0.],
+            [1., 1.], [1., 1.], [0.25, 0.5],
+            [0.5, 0.25], [0.5, 0.5], [0.5, 0.5],
+            [0.5, 0.75], [0.75, 0.5]
+            ]
+        
+        expected = star_poly
+        results, _ = concave_hull(points_2d,2.1)
+        assert results.contains(expected)
+        assert expected.contains(results)
+        # breakpoint()
+
+
+    def test_3_dimensional_regression(self):
         """
         Given a 3-dimensional data set, return an expected set of edges.
         """
@@ -90,47 +148,6 @@ class TestAlphashape(unittest.TestCase):
             (.5, .25, .5), (.5, .5, .25), (.75, .5, .5),
             (.5, .75, .5), (.5, .5, .75)
         ]
-        expected = {
-        }
-        expected_vertices = [
-            [0., 0., 0.], [0., 0., 1.], [0., 1., 0.],
-            [0., 1., 1.], [1., 0., 0.], [1., 0., 1.],
-            [1., 1., 0.], [1., 1., 1.], [0.25, 0.5, 0.5],
-            [0.5, 0.25, 0.5], [0.5, 0.5, 0.25], [0.5, 0.5, 0.75],
-            [0.5, 0.75, 0.5], [0.75, 0.5, 0.5]]
-        expected_faces = [
-            (13, 10, 6), (13, 9, 4), (6, 12, 13),
-            (13, 12, 7), (5, 11, 9), (8, 10, 0),
-            (3, 12, 8), (0, 10, 9), (5, 9, 13),
-            (12, 11, 7), (9, 10, 4), (8, 9, 1),
-            (12, 10, 2), (13, 11, 5), (1, 11, 8),
-            (4, 10, 13), (9, 11, 1), (2, 10, 8),
-            (8, 12, 2), (3, 11, 12), (0, 9, 8),
-            (7, 11, 13), (6, 10, 12), (8, 11, 3)]
-        results = alphashape(points_3d,2.1)
-        self.assertTrue(len(results.vertices) == len(expected_vertices))
-        self.assertTrue(len(points_3d) == len(expected_vertices))
-        self.assertTrue(len(results.faces) == len(expected_faces))
-        vertex_map = {i: expected_vertices.index(
-            list(vertex)) for i,vertex in enumerate(results.vertices)}
-        for edge in list(results.faces):
-           self.assertTrue(any([(
-               vertex_map[e[0]],
-               vertex_map[e[1]],
-               vertex_map[e[2]]) in expected_faces \
-                    for e in itertools.combinations(edge, r=len(edge))]))
-
-    # def test_3_dimensional_regression(self):
-    #     """
-    #     Given a 3-dimensional data set, return an expected set of edges.
-    #     """
-    #     points_3d = [
-    #         (0., 0., 0.), (0., 0., 1.), (0., 1., 0.),
-    #         (1., 0., 0.), (1., 1., 0.), (1., 0., 1.),
-    #         (0., 1., 1.), (1., 1., 1.), (.25, .5, .5),
-    #         (.5, .25, .5), (.5, .5, .25), (.75, .5, .5),
-    #         (.5, .75, .5), (.5, .5, .75)
-    #     ]
     #     expected = {
     #     }
     #     expected_vertices = [
